build: upgrade to TypeScript 6.0 and modernize build pipeline
### Build Pipeline

- Compile ESM only with `tsc`, generate CJS from ESM via SWC
- Delete `tsconfig.esm.json` and `tsconfig.cjs.json` — root `tsconfig.json`
  is now the single build config
- Add SWC ES2015 → ES5 syntax downleveling to maintain ES5 compatibility
- Run build scripts directly via `npx tsx` instead of compiling with `tsc`
- Simplify type declaration build step (single ESM target, CJS gets it via copy)

### TypeScript 6.0 Compatibility

- Set `target: "ES2015"` (ES5 emit removed in TS 7, SWC handles downleveling)
- Set `moduleResolution: "bundler"` (replaces deprecated `"node"`)
- Set explicit `module`, `esModuleInterop`, `rootDir`, `outDir`, `types`
  (required by TS 6.0 default changes)
- Remove deprecated `downlevelIteration`
- Remove `composite`, `tsBuildInfoFile`, `skipLibCheck`, and project references
  from ESLint tsconfig
- Simplify all tsconfigs by removing redundant/default options

### ESLint Configuration

- Move ESLint config to `eslint/eslint.config.ts` with a root re-export
  wrapper so VS Code uses the correct tsconfig for type-checking
- Add `tsconfigRootDir` to all parser options for correct path resolution
- Add `.d.ts` type override for `eslint-plugin-prefer-arrow-functions`
  upstream type incompatibility (typescript-eslint#9724)
- Delete unused `eslint-plugin-import.d.ts` (leftover from pre-import-x)
- Globally ignore `**/*.d.ts` from ESLint (type declarations don't need linting)

### Workarounds

- Inline minimal `BunPlugin` type in `src/bundler/bun.ts` to avoid
  `bun-types` WebAssembly type conflict with TS 6.0 (fix on bun main,
  not yet released in `@types/bun`)

### Other

- Update dependencies (TypeScript 6.0.2, eslint-plugin-unicorn 64, esbuild
  0.28, and other minor/patch updates)
- Update README badge to `min+gzip` label with dynamic inline size reference
- Normalize README badge casing to lowercase
