+++ package.json
@@ -117,10 +117,9 @@
     "dist"
   ],
   "scripts": {
-    "add-file-type-declaration": "node ./dist/esm/add-import-type.js && find ./dist -name 'add-import-type.*' -type f -delete",
     "benchmark": "npx tsx performance/benchmarks/run.ts",
     "benchmark-compare": "npx tsx performance/benchmarks/compare.ts",
-    "build": "npm run check-nvm-node-version && npm run prettier-write && npm run eslint-fix && rm -Rf ./dist && tsc -p tsconfig.cjs.json && tsc -p tsconfig.esm.json && echo '{ \"type\": \"commonjs\" }' > dist/cjs/package.json && tsc -p src/build-scripts/tsconfig.json && node dist/build-scripts/build.js && npm run test && npm run size",
+    "build": "npm run check-nvm-node-version && npm run prettier-write && npm run eslint-fix && rm -Rf ./dist && tsc && npx tsx src/build-scripts/build.ts && npm run test && npm run size",
     "check-nvm-node-version": "check-node-version --node $(node -p \"'>='+require('fs').readFileSync('.nvmrc','utf-8').trim()\")",
     "depcheck": "depcheck",
     "eslint-fix": "eslint --fix",
@@ -136,37 +135,38 @@
   },
   "devDependencies": {
     "@release-it/conventional-changelog": "10.0.6",
+    "@swc/core": "1.15.24",
     "@types/bun": "1.3.11",
     "@types/jest": "30.0.0",
-    "@types/node": "25.5.0",
+    "@types/node": "25.5.2",
     "check-node-version": "4.2.1",
     "depcheck": "1.4.7",
     "dotenv-cli": "11.0.0",
-    "esbuild": "0.27.4",
-    "eslint": "10.1.0",
+    "esbuild": "0.28.0",
+    "eslint": "10.2.0",
     "eslint-config-prettier": "10.1.8",
     "eslint-import-resolver-typescript": "4.4.4",
     "eslint-plugin-es-x": "9.6.0",
     "eslint-plugin-import-x": "4.16.2",
-    "eslint-plugin-jest": "29.15.0",
-    "eslint-plugin-package-json": "0.91.0",
+    "eslint-plugin-jest": "29.15.1",
+    "eslint-plugin-package-json": "0.91.1",
     "eslint-plugin-prefer-arrow-functions": "3.9.1",
     "eslint-plugin-prettier": "5.5.5",
     "eslint-plugin-tsdoc": "0.5.2",
-    "eslint-plugin-unicorn": "63.0.0",
+    "eslint-plugin-unicorn": "64.0.0",
     "husky": "9.1.7",
     "jest": "30.3.0",
     "jiti": "2.6.1",
     "jsonc-eslint-parser": "3.1.0",
-    "npm-check-updates": "19.6.5",
+    "npm-check-updates": "20.0.0",
     "prettier": "3.8.1",
     "release-it": "19.2.4",
-    "rollup": "4.60.0",
+    "rollup": "4.60.1",
     "terser": "5.46.1",
     "tinybench": "6.0.0",
-    "ts-jest": "29.4.6",
-    "typescript": "5.9.3",
-    "typescript-eslint": "8.57.1"
+    "ts-jest": "29.4.9",
+    "typescript": "6.0.2",
+    "typescript-eslint": "8.58.0"
   },
   "engines": {
     "node": "*"
