+++ performance/tsconfig.json
@@ -4,8 +4,8 @@
     "module": "esnext",
     "moduleResolution": "bundler",
     "strict": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
+    "types": ["node"]
   },
   "include": ["**/*.ts"],
   "exclude": ["benchmarks/.results/**", "size/.results/**", "snapshots/.snapshots/**"]
