+++ src/bundler/bun.ts
@@ -2,7 +2,22 @@ import { readFileSync } from 'node:fs'
 
 import { getProperties } from '..'
 
-import type { BunPlugin } from 'bun'
+/**
+ * Minimal BunPlugin type to avoid importing `bun-types` which has a WebAssembly
+ * type conflict with TypeScript 6.0+ (`wasm.d.ts` TS2403/TS2430).
+ *
+ * TODO: Replace with `import type { BunPlugin } from 'bun'` once `@types/bun` \> 1.3.11
+ * is released (fix is on main but not yet published).
+ */
+type BunPlugin = {
+  name: string
+  setup: (build: {
+    onLoad: (
+      options: { filter: RegExp },
+      callback: (loadArguments: { path: string }) => unknown
+    ) => void
+  }) => void
+}
 
 /**
  * Bun plugin for `.properties` files. Works with both `Bun.plugin` (runtime) and `Bun.build`
