--- tsconfig.esm.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "outDir": "dist/esm",
-    "moduleResolution": "node"
-  }
-}
